put the check if gift and child exist in their own function so i can call on them reguarly, added aditional testing

+++ loot.py
@@ -12,31 +12,32 @@ lootbag_db = '../lootbag.db'
 
 class Lootbag():
 
-  def terminalCatch(self, *terminalTuppleList):
+  def terminalCatch(self, terminalTuppleList):
+
 
     '''Filters through comands recieved in the command line '''
 
-    if len(terminalTuppleList[0]) < 2:
+    if len(terminalTuppleList) < 2:
       return "Please add an argument"
-    elif len(terminalTuppleList[0]) == 2:
-      if terminalTuppleList[0][1] == "ls":
-        return self.kidsWhoGetPressents(terminalTuppleList[0])
-      elif terminalTuppleList[0][1] == "help":
+    elif len(terminalTuppleList) == 2:
+      if terminalTuppleList[1] == "ls":
+        return self.kidsWhoGetPressents(terminalTuppleList)
+      elif terminalTuppleList[1] == "help":
         return self.help()
       else:
         return "Incorrect input"
-    elif len(terminalTuppleList[0]) == 3:
-      if terminalTuppleList[0][1] == 'delivered':
-        return self.toysDelivered(terminalTuppleList[0])
-      elif terminalTuppleList[0][1] == 'ls' and terminalTuppleList[0][2]:
-        return self.singleKidsPresents(terminalTuppleList[0])
+    elif len(terminalTuppleList) == 3:
+      if terminalTuppleList[1] == 'delivered':
+        return self.toysDelivered(terminalTuppleList)
+      elif terminalTuppleList[1] == 'ls' and terminalTuppleList[2]:
+        return self.singleKidsPresents(terminalTuppleList)
       else:
         return "Incorrect input"
-    elif len(terminalTuppleList[0]) == 4 :
-      if terminalTuppleList[0][1] == 'add':
-        return self.addToy(terminalTuppleList[0])
-      elif terminalTuppleList[0][1] == 'remove':
-        return self.removeToy(terminalTuppleList[0])
+    elif len(terminalTuppleList) == 4 :
+      if terminalTuppleList[1] == 'add':
+        return self.addToy(terminalTuppleList)
+      elif terminalTuppleList[1] == 'remove':
+        return self.removeToy(terminalTuppleList)
       else:
         return "Incorect input"
     else:
@@ -75,33 +76,62 @@ class Lootbag():
       except sqlite3.OperationalError as err:
                     print("oops", err)
 
-
-
-  def removeToy(self, termList):
-    '''Removes a toy from the child and from the database'''
+  def doesChildExist(self,childName):
+    '''Checks if a child is in the database'''
     with sqlite3.connect(lootbag_db) as conn:
       cursor = conn.cursor()
 
       cursor.execute(f'''SELECT ChildId
                           FROM Children
-                          WHERE Name = "{termList[2]}";''')
-      childname = cursor.fetchone()
-      print(childname)
+                          WHERE Name = "{childName}";''')
+      childId = cursor.fetchone()
+
+      return childId
+
+  def doesToyExist(self, toyName, childId):
+    '''Checks if a toy is in the database'''
     with sqlite3.connect(lootbag_db) as conn:
       cursor = conn.cursor()
 
       cursor.execute(f'''SELECT ToyId
                           FROM Toys
-                          WHERE ToyName = "{termList[3]}";''')
-      toyname = cursor.fetchone()
-      print(toyname)
+                          WHERE ToyName = "{toyName}"
+                          AND ChildId = {childId};''')
+      toyId = cursor.fetchone()
+      return toyId
+
+
+  def removeToy(self, termList):
+    '''Removes a toy from the child and from the database'''
+
+    # Calls funtion to check if a child exists
+    childId = self.doesChildExist(termList[2])
+    print("HERE", childId)
+    if childId == None:
+      print("That Child does not exist")
+      return childId
+
+
+    # Calls function to check if toy exist
+    toyId = self.doesToyExist(termList[3], childId[0])
+    if toyId == None:
+      print("That Toy does not exist")
+      return toyId
+
+
     with sqlite3.connect(lootbag_db) as conn:
       cursor = conn.cursor()
       cursor.execute(f'''DELETE FROM Toys
-                          WHERE ToyId = {toyname[0]}
-                          AND ChildId = {childname[0]}'''
-                         )
-    return "You removed :("
+                          WHERE ToyId = {toyId[0]}
+                          AND ChildId = {childId[0]}'''
+                        )
+    # Checks if the toy has been removed
+    isToyThere = self.doesToyExist(termList[3], childId[0])
+    if isToyThere == None:
+      print("Your toy is deleted")
+    else: print("Your toy could not be deleted")
+
+    return isToyThere
 
   def kidsWhoGetPressents(self, termList):
     '''List the kids who get presents'''
